Move file loading into new module summary_sample
